All group related functionality is now in groups.py
This includes model, schema, validation as well as group_permission and mailinlist hooks.

All other files have been adjusted to import correctly.
