+++ amivapi/schemas.py
@@ -3,25 +3,15 @@
 # license: AGPLv3, see LICENSE for details. In addition we strongly encourage
 #          you to buy us beer if we meet and you like the software.
 
-from sqlalchemy.ext.declarative import DeclarativeMeta
-
 from amivapi.utils import make_domain, EMAIL_REGEX
-from amivapi.db_utils import Base
 
-from events import Event, EventSignup
+from models import User, Session, File, StudyDocument, JobOffer, Purchase
 
 def get_domain():
     domain = {}
 
-    # generate from model
-    for model in Base._decl_class_registry.values():
-        if not isinstance(model, DeclarativeMeta):
-            continue
-
-        # Exlude resources already put in modules
-        if model in [Event, EventSignup]:
-            continue
-
+    # generate from models
+    for model in [User, Session, File, StudyDocument, JobOffer, Purchase]:
         domain.update(make_domain(model))
 
     # Now some modifications are required for each resource:
@@ -65,45 +55,6 @@ def get_domain():
 
     # POST will be handled by custom endpoint
     domain['sessions']['resource_methods'] = ['GET']
-
-    # /groups
-
-    # jsonschema validation for permissions
-    domain['groups']['schema']['permissions'].update({
-        'type': 'permissions_jsonschema'
-    })
-
-    # /groupaddresses
-    domain['groupaddresses']['schema']['group_id'].update({
-        'only_groups_you_moderate': True,
-        'unique_combination': ['email'],
-        'not_patchable': True,
-    })
-    domain['groupaddresses']['schema']['email'].update({
-        'regex': EMAIL_REGEX,
-        'unique_combination': ['group_id']})
-
-    # /groupforwards
-    domain['groupforwards']['schema']['group_id'].update({
-        'only_groups_you_moderate': True,
-        'unique_combination': ['email'],
-        'not_patchable': True,
-    })
-    domain['groupforwards']['schema']['email'].update({
-        'regex': EMAIL_REGEX,
-        'unique_combination': ['group_id']})
-
-    # /groupmembers
-
-    domain['groupmembers']['schema']['user_id'].update({
-        'only_self_enrollment_for_group': True,
-        'unique_combination': ['group_id']})
-    domain['groupmembers']['schema']['group_id'].update({
-        'self_enrollment_must_be_allowed': True,
-        'unique_combination': ['user_id']})
-
-    # Membership is not transferable -> remove PUT and PATCH
-    domain['groupmembers']['item_methods'] = ['GET', 'DELETE']
 
     # /files
