+++ amivapi/validation.py
@@ -10,17 +10,12 @@
     Also adds hooks to validate other input
 """
 
-import jsonschema
-
 from flask import g, current_app as app
 from werkzeug.datastructures import FileStorage
 from imghdr import what
 
 from eve_sqlalchemy.validation import ValidatorSQL
 from eve.utils import request_method
-
-from amivapi.utils import get_owner, get_class_for_resource
-from amivapi.group_permissions import create_group_permissions_jsonschema
 
 
 class ValidatorAMIV(ValidatorSQL):
@@ -114,108 +109,3 @@ class ValidatorAMIV(ValidatorSQL):
             self._error(field, "value already exists in the database in " +
                         "combination with values for: %s" %
                         unique_combination)
-
-    def _validate_only_self_enrollment_for_group(self, enabled, field, value):
-        """ Validates if the id can be used to enroll for a group,
-
-        Users can only sign up themselves
-        Moderators and admins can sign up everyone
-
-        :param enabled: boolean, validates nothing if set to false
-        :param field: field name.
-        :param value: field value.
-        """
-        if enabled:
-            # Get moderator id
-            group_id = self.document.get('group_id', None)
-            group = app.data.find_one("groups", None, id=group_id)
-
-            # If the group doesnt exist we dont have to do anything,
-            # The 'type' validator will generate an error anyway
-            if group is not None:
-                moderator_id = group["moderator_id"]
-
-                if not(g.resource_admin or (g.logged_in_user == value) or
-                       (g.logged_in_user == moderator_id)):
-                    self._error(field, "You can only enroll yourself. (%s: "
-                                "%s is yours)." % (field, g.logged_in_user))
-
-    def _validate_self_enrollment_must_be_allowed(self, enabled, field, value):
-        """ Validation for a group_id field in useraddressmembers
-        Validates if the group allows self enrollment.
-
-        Except group moderator and admins, they can ignore this
-
-        :param enabled: boolean, validates nothing if set to false
-        :param field: field name.
-        :param value: field value.
-        """
-        if enabled:
-            # Get moderator id
-            group = app.data.find_one("groups", None, id=value)
-
-            # If the group doesnt exist we dont have to do anything,
-            # The 'type' validator will generate an error anyway
-            if group is not None:
-                moderator_id = group["moderator_id"]
-                if not(g.resource_admin or (g.logged_in_user == moderator_id)
-                        or group["allow_self_enrollment"]):
-                    # This copies the validation error for the case this group
-                    # doesnt exist (since its hidden to the user)
-                    self._error(field,
-                                "value '%s' must exist in resource 'groups', "
-                                "field 'id'." % value)
-
-    def _validate_only_groups_you_moderate(self, enabled, field, value):
-        """ Validation for a group_id field in forwardaddresses
-        If you are not member or admin of the group you get the same
-        error as if the group wouldn't exist
-
-        If you are member, but not moderator, you will get a message that you
-        cannot enter this group_id
-
-        If you are moderator or have admin permissions it is alright.
-
-        :param enabled: boolean, validates nothing if set to false
-        :param field: field name.
-        :param value: field value.
-        """
-        if enabled:
-            # Get moderator id
-            group = app.data.find_one("groups", None, id=value)
-
-            # If the group doesnt exist we dont have to do anything,
-            # The 'type' validator will generate an error anyway
-            if group is not None:
-                if not g.resource_admin:
-                    moderator_id = group["moderator_id"]
-                    if not(g.logged_in_user == moderator_id):
-                        owners = get_owner(get_class_for_resource("groups"),
-                                           value)
-                        if g.logged_in_user in owners:
-                            self._error(field, "you are not the moderator of"
-                                        "this group.")
-                        else:
-                            # Not Member either
-                            # Copies the validation error for the case this
-                            # group doesnt exist (since its hidden to the user)
-                            self._error(field,
-                                        "value '%s' must exist in resource "
-                                        "'groups', field 'id'." % value)
-
-    def _validate_type_permissions_jsonschema(self, field, value):
-        """
-        Validates the jsonschema provided using the python jsonschema library
-
-        :param jsonschema: The jsonschema to use
-        :param field: field name.
-        :param value: field value.
-        """
-        schema = create_group_permissions_jsonschema(
-            app.config['DOMAIN'].keys())
-
-        try:
-            jsonschema.validate(value, schema)
-        except jsonschema.exceptions.ValidationError as v_error:
-            # Something was not according to schema
-            self._error(field, v_error.message)
